LayoutTests/media/track/track-css-stroke-cues-expected.txt


Test that style to all cues is applied correctly.
EVENT(canplaythrough)
EVENT(seeked)
EXPECTED (getComputedStyle(textTrackDisplayElement(video, 'cue')).webkitTextStrokeWidth == '4px') OK
EXPECTED (getComputedStyle(textTrackDisplayElement(video, 'cue')).webkitTextStrokeColor == 'rgb(0, 0, 0)') OK
END OF TEST
